requirements.txt
coloredlogs==15.0.1
fastapi==0.115.12
Markdown==3.8
pdfplumber==0.11.6
pytest==8.3.5
python-dotenv==1.1.0
python_docx==1.1.2
PyYAML==6.0.2
SQLAlchemy==2.0.40
starlette==0.46.2
# textract==1.6.3
uvicorn==0.34.2
pydantic==2.11.4

# LangChain and Multi-Agent System Dependencies
langchain-core==0.3.65
langchain-openai==0.3.22
langchain-google-genai
langchain-groq
langchain_anthropic==0.3.15
langgraph==0.4.8

# Async Task Processing and Real-time Updates
celery==5.3.6
redis==5.0.8
websockets==12.0
email-validator==2.1.1

# Email Support
aiosmtplib==3.0.1

# Authentication and JWT
PyJWT==2.10.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# Google OAuth
google-auth==2.36.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0